Rework configuration.
Most of the configuration is now stored in hypes.
If configuration is not defined in the hype file,
system defaults, controlled by env_variabels are used.
If there are no env_variabels set TV defaults will be used.

Additionally this commit improves pep8 compatibility.

Additionally a new switch --gpus is included to improve cluster support.