Tv analyze (#34)
* Create core.py and move some functions there.

* Implement first Version of tv-analyze.

This Version loads the graph from the logdir and runs a do_eval() on the validation data.

* Update _write_evaluation_to_summary to make it write all evaluation results to tensorboard.

do_eval now outputs a tuple of lists representing the results of the evaluated metrics. Those metrics are now written to Tensorboard using tf.summary.

* Improve code quality by many minor changes

* Remove pylint comments
* Add intersphinx for docs
* Fix docstrings
* Remove unused module imports
* Remove code duplication
* Use only two epochs for cifar10_minimal for testing purposes

Mention tv-analyze in docs and add analyze and core to the documentation.